expose fail matchers for testing custom matchers
