+++ spec/rspec/matchers_spec.rb
@@ -13,7 +13,7 @@ RSpec.describe RSpec::Matchers do
     in_sub_process do
       main.instance_eval do
         include RSpec::Matchers
-        include FailMatchers
+        include RSpec::Matchers::FailMatchers
 
         expect(3).to eq(3)
         expect(3).to be_odd
