+++ spec/support/matchers.rb
@@ -1,3 +1,5 @@
+require 'rspec/matchers/fail_matchers'
+
 RSpec::Matchers.define :include_method do |expected|
   match do |actual|
     actual.map { |m| m.to_s }.include?(expected.to_s)
@@ -32,25 +34,9 @@ RSpec::Matchers.define :have_string_length do |expected|
   end
 end
 
-module FailMatchers
-  def fail
-    raise_error(RSpec::Expectations::ExpectationNotMetError)
-  end
-
-  def fail_with(message)
-    raise_error(RSpec::Expectations::ExpectationNotMetError, message)
-  end
-
-  def fail_including(snippet)
-    raise_error(
-      RSpec::Expectations::ExpectationNotMetError,
-      a_string_including(snippet)
-    )
-  end
-end
 
 RSpec.configure do |config|
-  config.include FailMatchers
+  config.include RSpec::Matchers::FailMatchers
 end
 
 RSpec::Matchers.define_negated_matcher :a_string_excluding, :a_string_including
